+++ nautobot_device42_sync/diffsync/d42utils.py
@@ -113,22 +113,20 @@ def get_netmiko_platform(network_os: str) -> str:
     return network_os
 
 
-def find_device_role_from_tags(diffsync, tag_list: List[str]) -> str or bool:
+def find_device_role_from_tags(diffsync, tag_list: List[str]) -> str:
     """Determine a Device role based upon a Tag matching the `role_prepend` setting.
 
     Args:
         tag_list (List[str]): List of Tags as strings to search.
 
     Returns:
-        DEFAULTS["device_role"]: The Default device role defined in plugin settings.
+        str: The Default device role defined in plugin settings.
     """
-    if not PLUGIN_CFG.get("role_prepend"):
-        print("You must have the `role_prepend` setting configured.")
-        raise MissingConfigSetting(setting="role_prepend")
     _prepend = PLUGIN_CFG.get("role_prepend")
-    for _tag in tag_list:
-        if re.search(_prepend, _tag):
-            return re.sub(_prepend, "", _tag)
+    if _prepend:
+        for _tag in tag_list:
+            if re.search(_prepend, _tag):
+                return re.sub(_prepend, "", _tag)
     return DEFAULTS.get("device_role")
 
 
@@ -399,7 +397,7 @@ class Device42API:
                 "notes": _cf["notes"],
             }
             _fields[f"{_cf['network']}/{_cf['mask_bits']}"].append(_field)
-        
+
         for _, cfields in _fields.items():
             cfields = sorted(cfields, key=lambda d: d["key"])